+++ src/types/apiClient.ts
@@ -4,6 +4,7 @@ import axios from 'axios';
 interface AxiosErrorLike {
   response?: {
     data?: unknown;
+    status?: number;
   };
 }
 
@@ -15,17 +16,36 @@ const isAxiosError = (error: unknown): error is AxiosErrorLike => {
 const metaEnv =
   (import.meta as unknown as { env?: Record<string, string | undefined> })
     .env ?? {};
-const API_SERVER = metaEnv.VITE_API_SERVER;
+const API_SERVER =
+  metaEnv.VITE_API_SERVER || 'https://fesp-api.koyeb.app/market';
 
-// ✅ Axios 인스턴스 생성
-const api = axios.create({
+// ✅ Axios 인스턴스 생성 (client-id 및 인증 헤더 포함)
+export const api = axios.create({
   baseURL: API_SERVER,
   withCredentials: false,
   headers: {
     'Content-Type': 'application/json',
-    'client-id': metaEnv.VITE_CLIENT_ID ?? '',
+    'client-id': metaEnv.VITE_CLIENT_ID || 'brunch',
   },
 });
+
+// FormData를 사용하는 요청을 위한 인터셉터 추가
+api.interceptors.request.use(
+  config => {
+    // FormData를 사용하는 경우 (multipart/form-data)
+    // 기본 Content-Type 헤더를 제거하여 axios가 자동으로 boundary를 포함한 헤더를 설정하도록 함
+    if (config.data instanceof FormData) {
+      // 헤더에서 Content-Type 제거 (axios가 자동 설정)
+      if (config.headers) {
+        delete config.headers['Content-Type'];
+      }
+    }
+    return config;
+  },
+  error => {
+    return Promise.reject(error);
+  },
+);
 
 const LOCAL_USERS_STORAGE_KEY = 'vanilla:signup:users';
 
@@ -34,9 +54,10 @@ export type ProviderVariant = 'local' | 'kakao';
 export interface LocalRegisteredUser {
   email: string;
   nickname: string;
-  provider: ProviderVariant | string;
+  image?: string;
   type: string;
   password?: string;
+  provider?: ProviderVariant;
 }
 
 const normalizeValue = (value: string): string => value.trim().toLowerCase();
@@ -49,8 +70,9 @@ const isValidLocalUser = (item: unknown): item is LocalRegisteredUser => {
   return (
     typeof candidate.email === 'string' &&
     typeof candidate.nickname === 'string' &&
-    typeof candidate.provider === 'string' &&
-    typeof candidate.type === 'string'
+    typeof candidate.type === 'string' &&
+    (typeof candidate.provider === 'string' ||
+      typeof candidate.provider === 'undefined')
   );
 };
 
@@ -63,7 +85,7 @@ export const loadLocalRegisteredUsers = (): LocalRegisteredUser[] => {
     if (!raw) {
       return [];
     }
-    const parsed = JSON.parse(raw);
+    const parsed = JSON.parse(raw) as unknown;
     if (!Array.isArray(parsed)) {
       return [];
     }
@@ -91,7 +113,6 @@ const storeLocalRegisteredUsers = (users: LocalRegisteredUser[]): void => {
         users.map(user => ({
           email: normalizeValue(user.email),
           nickname: normalizeValue(user.nickname),
-          provider: user.provider,
           type: user.type,
           password: user.password,
         })),
@@ -107,7 +128,7 @@ export const addLocalRegisteredUser = (user: LocalRegisteredUser): void => {
   current.push({
     email: normalizeValue(user.email),
     nickname: normalizeValue(user.nickname),
-    provider: user.provider,
+    image: user.image,
     type: user.type,
     password: user.password,
   });
@@ -116,19 +137,19 @@ export const addLocalRegisteredUser = (user: LocalRegisteredUser): void => {
 
 export const isEmailRegisteredLocally = (email: string): boolean => {
   const normalized = normalizeValue(email);
-  return loadLocalRegisteredUsers().some(user => user.email === normalized);
+  return loadLocalRegisteredUsers().some(u => u.email === normalized);
 };
 
 export const isNicknameRegisteredLocally = (nickname: string): boolean => {
   const normalized = normalizeValue(nickname);
-  return loadLocalRegisteredUsers().some(user => user.nickname === normalized);
+  return loadLocalRegisteredUsers().some(u => u.nickname === normalized);
 };
 
 export const findLocalRegisteredUser = (
   email: string,
 ): LocalRegisteredUser | undefined => {
   const normalized = normalizeValue(email);
-  return loadLocalRegisteredUsers().find(user => user.email === normalized);
+  return loadLocalRegisteredUsers().find(u => u.email === normalized);
 };
 
 export interface ApiResponse<T> {
@@ -169,7 +190,6 @@ export interface User {
   email: string;
   password?: string;
   name?: string;
-  provider?: string;
   image?: string;
   type?: string;
   extra?: Record<string, unknown>;
@@ -184,7 +204,6 @@ export const loginUser = async (payload: {
 }): Promise<ApiResponse<User>> => {
   try {
     const { data } = await api.post<ApiResponse<User>>('/users/login', payload);
-
     // 항상 200이므로 throw 없음
     return data;
   } catch (err) {
@@ -245,7 +264,6 @@ export const getKakaoUserInfo = async (
   }
 };
 
-
 // ✅ 응답 타입 정의
 export interface ApiItemResponse<T> {
   ok: boolean;
@@ -266,15 +284,19 @@ export const registerUser = async (
   userData: User,
 ): Promise<ApiItemResponse<User>> => {
   try {
-    const extraPayload = { ...(userData.extra ?? {}) };
-    if (!extraPayload.providerAccountId) delete extraPayload.providerAccountId;
+    const extraPayload = { ...(userData.extra ?? {}) } as Record<
+      string,
+      unknown
+    >;
+    if (!('providerAccountId' in extraPayload)) {
+      // nothing to send
+    }
 
     const payload = {
       email: userData.email,
       password: userData.password,
       name: userData.name,
       type: userData.type ?? 'user',
-      loginType: userData.provider ?? 'local',
       ...(userData.image && { image: userData.image }),
       ...(Object.keys(extraPayload).length > 0 && { extra: extraPayload }),
     };
@@ -301,4 +323,87 @@ export const getUserById = async (
 ): Promise<ApiItemResponse<User>> => {
   const { data } = await api.get<ApiItemResponse<User>>(`/users/${id}`);
   return data;
+};
+
+export const getUserByPath = async (
+  path: string,
+): Promise<ApiItemResponse<User>> => {
+  const normalizedPath = path.replace(/^\/+/, '');
+  const { data } = await api.get<ApiItemResponse<User>>(
+    `/users/${normalizedPath}`,
+  );
+  return data;
+};
+
+export const getUserByEmail = async (
+  email: string,
+): Promise<ApiItemResponse<User>> => {
+  const normalized = normalizeValue(email);
+  const { data } = await api.get<ApiItemResponse<User>>(`/users/email`, {
+    params: { email: normalized },
+  });
+  return data;
+};
+
+export const isEmailRegisteredInDb = async (
+  email: string,
+): Promise<boolean> => {
+  const normalized = normalizeValue(email);
+  try {
+    const response = await getUserByEmail(normalized);
+    const user = response.data ?? response.item;
+    return Boolean(user);
+  } catch (error) {
+    if (isAxiosError(error)) {
+      const status = error.response?.status;
+      if (status === 404) {
+        return false;
+      }
+      // 서버가 409(Conflict)로 중복을 알리는 경우 → 사용 중인 이메일
+      if (status === 409) {
+        return true;
+      }
+    }
+    console.error(
+      '[apiClient] Failed to check server email duplication:',
+      error,
+    );
+    throw error;
+  }
+};
+
+// ✅ 닉네임으로 사용자 조회 (/users/name?name=...)
+export const getUserByName = async (
+  name: string,
+): Promise<ApiItemResponse<User>> => {
+  const normalized = normalizeValue(name);
+  const { data } = await api.get<ApiItemResponse<User>>(`/users/name`, {
+    params: { name: normalized },
+  });
+  return data;
+};
+
+// ✅ DB 닉네임 중복 여부 확인
+export const isNameRegisteredInDb = async (
+  name: string,
+): Promise<boolean> => {
+  const normalized = name.trim();
+  if (!normalized) return false;
+  try {
+    const response = await getUserByName(normalized);
+    const user = response.data ?? response.item;
+    return Boolean(user);
+  } catch (error) {
+    if (isAxiosError(error)) {
+      const status = error.response?.status;
+      if (status === 404) {
+        return false;
+      }
+    }
+    console.error(
+      '[apiClient] Failed to check server name duplication:',
+      error,
+    );
+    throw error;
+  }
 };
